Some cleanup and new process ID values from threads
