+++ prototypes/conditionals/conditional.c
