+++ prototypes/errors/errors.c
