+++ prototypes/fileIO/write_file.c
