+++ prototypes/syslog/syslog.c
